requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
defusedxml==0.7.1
diff-match-patch==20230430
Django
django-appconf
django-browser-reload
django-components
django-compressor
django-environ
django-import-export
django-resized
django-tailwind
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.6
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
mdurl==0.1.2
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
psycopg2-binary==2.9.9
Pygments==2.17.2
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
rich==13.7.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
types-python-dateutil==2.8.19.14
tzdata==2023.4
urllib3==2.1.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
